Add files via upload
this is the first commit. you ll find the first functions to initialize the grid for the 2048 game.
there s a small problem where in the beginning, you ll always get the same number (either 2 or 4) twice in the starting squares, when it should be possible to get a 2 and a 4.