Add files via upload
you can now use your keyboard to play the game. i still have to fix a slight problem in the fusion of cells since it does it all the way through and not just once. also i ll try to add the restart command, and the cancel move function later on.